Create Interactive Map

Generates an interactive map visualizing conservation footprints and additional layers specified by the user.

Usage

create_interactive_map(
  solution_idx = NULL,
  dir = ".",
  run_id = "optimTFE",
  spatial = NULL,
  map_tiles = "Esri.WorldTopoMap",
  spp_layers = NULL,
  auxiliary_layers = NULL,
  auxiliary_palette = "Set1",
  solution_color = "green",
  solution_opacity = 0.5,
  html_out = NULL
)

Arguments

solution_idx

Integer, index of the solution to visualize.

dir

Character, directory containing the optimTFE output. Default is current directory.

run_id

Character, identifier for the optimTFE run. Default is "optimTFE".

spatial

SF or Spatial object, containing planning unit polygons with unit_id field.

map_tiles

Character, name of the leaflet provider background tiles. Default is "Esri.WorldTopoMap". Set to NULL to use no background map tiles.

spp_layers

Character vector, names of species to add as separate overlay layers.

auxiliary_layers

Named list of paths to additional spatial files (or pre-loaded sf objects) to add as overlay layers.

auxiliary_palette

RColorBrewer palette to use for auxiliary layers (default is "Set1").

solution_color

Character, color for the solution layer. Default is "green".

solution_opacity

Numeric, opacity for the solution layer. Default is 0.5.

html_out

Character, file path to save a self contained, sharable HTML file.

Value

A leaflet map object.

See also