Launch the Criteria Explorer Shiny Application

Launch the Criteria Explorer Shiny Application

Usage

criteria_explorer(
  spatial = NULL,
  data = NULL,
  dir = ".",
  run_id = "optimTFE",
  criteria_include = NULL,
  criteria_presets = NULL,
  map_tiles = "Esri.WorldTopoMap",
  auxiliary_layers = NULL
)

Arguments

spatial

Either a file path to spatial data (e.g., GeoPackage, Shapefile) or an `sf` object containing polygon geometries. The first column should contain unique unit identifiers that match those in the solutions data.

data

Either a file path to a CSV file containing solution summary data or a data.frame. Must include a 'units' column with JSON-formatted unit lists. If NULL, will attempt to load from `file.path(dir, run_id, paste0(run_id, ".summary.csv"))`.

dir

Character string specifying the directory containing optimization results. Defaults to current working directory (".").

run_id

Character string identifying the optimization run. Used to construct file paths when `data` is NULL. Defaults to "optimTFE".

criteria_include

Character vector specifying column names to include as selection criteria. If NULL (the default), all columns from the input data will be used.

criteria_presets

Named list specifying a set of selection criteria to be preselected when the application is launched. Must be a named list where the names specific the criteria to be preselected. Each element should be a list with 2 elements, weight (0>1) and descending (T/F). For example: `list(accessibility = list(weight = 0.5, descending = T))`

map_tiles

Character, name of the leaflet provider tiles. Default is "Esri. WorldTopoMap". Set to NULL to use no background map tiles.

auxiliary_layers

Named list, paths to additional spatial files to add as overlay layers.

Details

The application provides:

Interactive filtering of solutions based on summary statistics

Spatial visualization of solution footprints

Histogram displays of solution criteria

Ranking and comparison tools

See also

Examples

if (FALSE) { # \dontrun{
# Launch with default settings
criteria_explorer(
  spatial = "path/to/spatial.gpkg",
  run_id = "optimTFE" # optimTFE run id
)
} # }