Generate summary statistics for optimization solutions

This function calculates various summary statistics for optimization solutions, including unit counts, suitability metrics, species richness, and spatial metrics (when spatial data is provided).

Usage

generate_summary(
  data = NULL,
  out_dir = ".",
  run_id = "optimTFE",
  suitability_mx = suitability_mx,
  unit_ids = unit_ids,
  spatial = spatial,
  spatial_projection = NULL,
  min_units_plot = file.path(out_dir, run_id, "n_units.pdf"),
  summary_out = file.path(out_dir, run_id, "summary.csv"),
  return_df = FALSE
)

Arguments

data

Dataframe containing solution data. If NULL, data will be loaded from the specified output directory.

out_dir

Character string specifying the output directory path. Default is ".".

run_id

Character string specifying the run identifier. Default is "optimTFE".

suitability_mx

Matrix of suitability values with unit IDs as row names and species as column names.

unit_ids

Vector of unit IDs (must be in the same order as the rows of suitability_mx).

spatial

Spatial data frame (sf object) containing geometries for planning units.

spatial_projection

CRS object or string for transforming spatial data to an equal area projection if needed.

min_units_plot

Character string specifying the output path for the minimum units plot description (default = `file.path(out_dir, run_id, "n_units.pdf")`).

summary_out

Character string specifying the output path for the summary data frame (default = `file.path(out_dir, run_id, "summary.csv")`).

return_df

Logical value indicating whether to return the summary data frame.

Value

A data frame with summary statistics for each solution, including number of units, mean/median suitability, species richness metrics, and spatial metrics (area and perimeter) if spatial data is provided.